move cache and reuse global cache
